requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
Werkzeug==3.0.0
openpyxl==3.1.5
Pillow==12.0.0
python-dotenv==1.0.1
reportlab==4.4.4
certifi
